experimental/llm_truss/llama2-7b-sample/requirements.txt
torch==2.0.1
peft
sentencepiece
accelerate
bitsandbytes
google-cloud-secret-manager==2.16.1
einops
scipy
protobuf
git+https://github.com/huggingface/transformers.git